Adding Image of working UI
The center image is of the original painting. 

The image to the left is a values filter showing dark mid and light tones. Generally, you want the greatest contrast to exist at the focal point.

The image to the right is a light filter showing what kinds of light are most prominent in the painting. As is my common habit, I tend to use a lot of red and green in my paintings. This filter reminds me to consider using more cool colors to contrast my warm tones.

Other Educational Filters are also available.

**I painted the original image